#32 go파일명 폴더명과 동일하게 변경 StripPrefix -> 요청URL의 경로에서 parameter로 받은 접두사를 제거한 후 parameter의 핸들러를 호출하게 한다.
